update for ConTED

+++ trustlab/models.py
@@ -1,14 +1,15 @@
-from trustlab_host.models import Scenario
-from django.db import models
-import re
 import importlib
 import importlib.util
 import inspect
-from os import listdir, mkdir
-from os.path import isfile, exists, isdir
 import pprint
-from trustlab.lab.config import SCENARIO_PATH, SCENARIO_PACKAGE, RESULT_PATH
+import re
 import traceback
+from django.db import models
+from os import listdir, mkdir
+from os.path import isfile, exists, isdir, getsize, basename
+from pathlib import Path
+from trustlab.lab.config import SCENARIO_PATH, SCENARIO_PACKAGE, RESULT_PATH, SCENARIO_LARGE_SIZE, SCENARIO_CATEGORY_SORT
+from trustlab_host.models import Scenario
 
 
 class Supervisor(models.Model):
@@ -22,24 +23,22 @@ class ObjectFactory:
     Generic Class for Object Factories loading and saving objects in a DSL (.py) file.
     """
     @staticmethod
-    def load_object(file_package, object_package, object_class_name, object_args):
+    def load_object(import_package, object_class_name, object_args, lazy_args=None):
         """
         Imports the DSL file of an given object and initiates it.
 
-        :param file_package: package name of the DSL file.
-        :type file_package: str
-        :param object_package: package name of the object's class.
-        :type object_package: str
+        :param import_package: python package path of the DSL file.
+        :type import_package: str
         :param object_class_name: the name of the object's class.
         :type object_class_name: str
         :param object_args: All the parameters of the to initiate object
         :type object_args: inspect.FullArgSpec
+        :param lazy_args: List of arguments for the lazy load of too large files. Default is None.
+        :type lazy_args: list
         :return: the initiated object to be loaded
         :rtype: Any
         :raises AttributeError: One or more mandatory attribute was not found in object's DSL file.
         """
-        # python package path
-        import_package = f"{object_package}.{file_package}"
         # ensure package is accessible
         implementation_spec = importlib.util.find_spec(import_package)
         if implementation_spec is not None:
@@ -55,28 +54,36 @@ class ObjectFactory:
             all_args = [a.upper() for a in object_args.args[1:]]
             # check if all mandatory args are in scenario config
             if all(hasattr(object_config_module, attr) for attr in mandatory_args):
-                object_attrs = []
-                for attr in all_args:
-                    # check if attr is in config as optional ones may not be present with allowance
-                    if hasattr(object_config_module, attr):
-                        object_attrs.append(getattr(object_config_module, attr))
+                object_attrs = {}
+                if not lazy_args:
+                    for attr in all_args:
+                        # check if attr is in config as optional ones may not be present with allowance
+                        if hasattr(object_config_module, attr):
+                            object_attrs[attr.lower()] = getattr(object_config_module, attr)
+                else:
+                    for attr in lazy_args:
+                        # check if attr is in config as optional ones may not be present with allowance
+                        if hasattr(object_config_module, attr):
+                            object_attrs[attr.lower()] = getattr(object_config_module, attr)
                 # get object's class by name, requires class to be in global spec thus imported
                 object_class = globals()[object_class_name]
-                # add all attrs which are in config to scenario object
-                obj = object_class(*object_attrs)
+                # add all attrs which are in config to object
+                obj = object_class(**object_attrs)
                 return obj
             raise AttributeError("One or more mandatory attribute was not found in object's DSL file.")
 
     def save_object(self, obj, object_args, file_path, file_exists=False):
         """
-        Saves an object to a DSL file.
+        Saves an object to a DSL file. Double new lines (empty line) within value definition of variable causes error
+        as script sees \n\n as indicator for variable end.
+        Only writes into files if object or its string representation has changed since loading.
 
         :param obj: the object to be saved.
         :type obj: Any
         :param object_args: All the parameters of the to initiate object
         :type object_args: inspect.FullArgSpec
         :param file_path: Full path to the object's DSL file.
-        :type file_path: pathlib.Path
+        :type file_path: Path
         :param file_exists: whether the object's DSL file already exists.
         :type file_exists: bool
         :return: the initiated object to be loaded
@@ -88,26 +95,42 @@ class ObjectFactory:
         if file_exists:
             with open(file_path, 'r+') as object_file:
                 # read in file
-                object_data = object_file.read()
+                # add newline feed for parsing with double new lines as delimiter for value end
+                object_data = object_file.read() + '\n'
                 # exchange all args which are in config file data
                 for arg in all_args:
                     # create regex to find argument with value.
-                    replacement = re.compile(arg + r' = .+?\n\n', re.DOTALL)  # variables ends with double new lines
+                    replacement = re.compile(arg + r' = .+?\n\n', re.DOTALL)  # variables end with double new lines
                     value = self.stringify_arg_value(obj, arg)
-                    if re.search(replacement, object_data):
-                        # substitute current value in config_data. Double new lines are added to help parsing
-                        object_data = replacement.sub(f"{arg} = {value}\n\n", object_data)
+                    match = re.search(replacement, object_data)
+                    arg_value_str = f"{arg} = {value}\n\n"
+                    object_changed = False
+                    if match and match.group() != arg_value_str:
+                        # only substitute value if object changed
+                        # substitute current value in config_data.
+                        object_data = replacement.sub(arg_value_str, object_data)
+                        # set object_changed to true to indicate file writing later on
+                        object_changed = True
+                    elif match and match.group() == arg_value_str:
+                        # not changing anything if file does not need a change for this arg
+                        pass
                     else:
+                        # as object added obligatory argument, file needs to be written again
+                        object_changed = True
                         # get position of last non whitespace char in config data
                         position = object_data.rfind(next((char for char in reversed(object_data) if char != "\n"
                                                            and char != "\t" and char != " "))) + 1
-                        arg_value = f"\n\n{arg} = {value}"  # Double new lines are added to help parsing
+                        arg_value = f"\n\n{arg} = {value}"  # Double new lines are added to support parsing
                         # append argument configuration at position -> end of file + whitespace tail
                         object_data = object_data[:position] + arg_value + object_data[position:]
-                # jump back to begin of file and write new data
-                object_file.seek(0)
-                object_file.write(object_data)
-                object_file.truncate()
+                # only write anything if object changed after load
+                if object_changed:
+                    # delete last new line feed to format in PEP8 style
+                    object_data = object_data[:-1]
+                    # jump back to begin of file and write new data
+                    object_file.seek(0)
+                    object_file.write(object_data)
+                    object_file.truncate()
         else:
             with open(file_path, 'w+') as object_file:
                 print('"""', file=object_file)
@@ -115,8 +138,8 @@ class ObjectFactory:
                 print('"""\n\n', file=object_file)
                 for arg in all_args:
                     value = self.stringify_arg_value(obj, arg)
-                    print(f"{arg} = {value}\n", file=object_file)  # Double new lines are added to help parsing
-                print("\n\n\n", file=object_file) # 4 new lines at end of file
+                    # Double new lines are added to support parsing
+                    print(f"{arg} = {value}\n\n", file=object_file)
 
     @staticmethod
     def stringify_arg_value(obj, arg):
@@ -126,6 +149,9 @@ class ObjectFactory:
         value = getattr(obj, arg.lower())
         # Prettifying the value for better human readability.
         value_prettified = pprint.pformat(value)
+        # delete leading and trailing brackets at string prettifying AND add backslash to newline feed
+        if type(value) is str and value_prettified.startswith('('):
+            value_prettified = value_prettified.replace('\n', '\\\n')[1:-1]
         return value_prettified
 
     def __init__(self):
@@ -137,18 +163,32 @@ class ScenarioFactory(ObjectFactory):
         """
         Loads all scenarios saved /trustlab/lab/scenarios with dynamic read of parameters from Scenario.__init__.
 
-        :return: scenarios initialized as Scenario objects
+        :return: scenarios initialized as Scenario objects or tuple representation from lazy load
         :rtype: list
         """
         scenarios = []
-        scenario_file_names = [file for file in listdir(self.scenario_path)
-                               if isfile(self.scenario_path / file) and file.endswith("_scenario.py")]
+        scenario_files = self.get_scenario_files()
         # get all parameters of scenario init
-        scenario_args = inspect.getfullargspec(Scenario.__init__)
-        for file_name in scenario_file_names:
-            file_package = file_name.split(".")[0]
+        scenario_args = inspect.getfullargspec(Scenario.scenario_args)
+        # only take name and description as more is not required for lazy load
+        scenario_lazy_args = ['NAME', 'DESCRIPTION']
+        for file_name, file_package in scenario_files:
+            file_size = getsize(self.scenario_path / file_name)
             try:
-                scenario = self.load_object(file_package, SCENARIO_PACKAGE, "Scenario", scenario_args)
+                if not self.large_file_size or (self.large_file_size and file_size < self.large_file_size):
+                    scenario = self.load_object(f"{SCENARIO_PACKAGE}.{file_package}", "Scenario", scenario_args)
+                else:
+                    scenario = self.load_object(f"{SCENARIO_PACKAGE}.{file_package}", "Scenario", scenario_args,
+                                                scenario_lazy_args)
+                    if self.large_file_size > 999999:
+                        file_size_str = f'{file_size / 1000000} MB'
+                        large_size_str = f'{self.large_file_size / 1000000} MB'
+                    else:
+                        file_size_str = f'{file_size / 1000} KB'
+                        large_size_str = f'{self.large_file_size / 1000} KB'
+                    scenario.lazy_note = f'This scenario file exceeded with its file size of {file_size_str} ' \
+                                         f'the file size limit of {large_size_str}. Thus, the scenario was lazy ' \
+                                         f'loaded and will only include its description.'
             except (ValueError, AttributeError, TypeError, ModuleNotFoundError, SyntaxError):
                 print(f'Error at Scenario file @{file_name}:')
                 traceback.print_exc()
@@ -200,14 +240,56 @@ class ScenarioFactory(ObjectFactory):
             if scenario.any_agents_use_metric('content_trust.topic'):
                 scenario.topics = scenario.agents_with_metric('content_trust.topic')
 
-    def __init__(self):
+    def get_scenario_files(self):
+        """
+        To categorize scenarios, this method scans for scenarios in direct subdirs of the scenario dir.
+
+        :return: All scenario files as tuples with their package name.
+        :rtype: list
+        """
+        scenario_file_names = [file for file in listdir(self.scenario_path)
+                               if isfile(self.scenario_path / file) and file.endswith("_scenario.py")]
+        subpackages = [self.scenario_path / sub for sub in listdir(self.scenario_path)
+                       if isdir(self.scenario_path / sub) and '__init__.py' in listdir(self.scenario_path / sub)]
+        for subpackage in subpackages:
+            subpackage_scenarios = [f'{basename(subpackage)}/{file}' for file in listdir(subpackage)
+                                    if isfile(subpackage / file) and file.endswith("_scenario.py")]
+            scenario_file_names += subpackage_scenarios
+        scenario_files = [(file, file.split(".")[0].replace('/', '.')) for file in scenario_file_names]
+        return scenario_files
+
+    def get_scenarios_in_categories(self):
+        """
+        Creates a dict of categories as keys and list of scenario names as values to visualize in scenario.hmtl card.
+
+        :return: Dictionary items of categories as keys, and list of scenario's names related to the scenario.
+        :rtype: list
+        """
+        scenario_categories = {}
+        for scenario in self.scenarios:
+            if '/' in scenario.file_name:
+                category = scenario.file_name.split('/')[0]
+            else:
+                category = 'Misc'
+            if category not in scenario_categories.keys():
+                scenario_categories[category] = []
+            scenario_categories[category].append(scenario.name)
+        for key in scenario_categories.keys():
+            scenario_categories[key] = sorted(scenario_categories[key])
+        # sort by categories
+        category_sort = SCENARIO_CATEGORY_SORT + sorted([c for c in scenario_categories.keys() if c not in SCENARIO_CATEGORY_SORT and c != 'Misc']) + ['Misc']
+        index_map = {v: i for i, v in enumerate(category_sort)}
+        return sorted(scenario_categories.items(), key=lambda pair: index_map[pair[0]])
+
+    def __init__(self, lazy_load=False):
         super().__init__()
         self.scenario_path = SCENARIO_PATH
+        self.large_file_size = SCENARIO_LARGE_SIZE if lazy_load else None
         self.scenarios = self.load_scenarios()
-        self.init_scenario_number = len(self.scenarios)
 
     def __del__(self):
-        self.save_scenarios()
+        if not self.large_file_size:
+            self.save_scenarios()
 
 
 class ScenarioResult:
@@ -286,4 +368,3 @@ class ResultFactory:
         self.result_path = RESULT_PATH
         if not exists(RESULT_PATH) or not isdir(RESULT_PATH):
             mkdir(RESULT_PATH)
-
